[is_palindrome] Types collisions updates

+++ palindrome_integer/0-is_palindrome.c
@@ -1,3 +1,4 @@
+#include <stdio.h>
 #include "palindrome.h"
 
 /**
@@ -9,11 +10,11 @@
  */
 int is_palindrome(unsigned long n)
 {
-	unsigned int _n = n;
-	unsigned int numHighUnit = 0;
-	unsigned int resultat = 0;
-	unsigned int operand = 1;
-
+	unsigned long _n = n;
+	unsigned long numHighUnit = 0;
+	unsigned long resultat = 0;
+	unsigned long operand = 1;
+	
 	/**
 	 * STEP 1: Search the highest pow of the number
 	 */
@@ -23,14 +24,13 @@ int is_palindrome(unsigned long n)
 		_n /= 10;
 	}
 	_n = n;
-
-	for (unsigned int i = 0; i < numHighUnit; i++)
+	
+	for (unsigned long i = 0; i < numHighUnit; i++)
 		operand *= 10;
-
 	/**
 	 * STEP 2: Build a new number with modulus of each pow level
 	 */
-	for (unsigned int i = numHighUnit; i > 0; i--)
+	for (unsigned long i = numHighUnit; i > 0; i--)
 	{
 		resultat += ((n % 10) * operand);
 		operand /= 10;
